Morris Inorder Traversal of a Binary Tree
Approach:

When we are currently at a node, the following cases can arise:

Case 1: When the current node has no left subtree. In this scenario, there is nothing to be traversed on the left side, so we print the current node's value and move to the right of the current node.
Case 2: When there is a left subtree the right-most child of this left subtree is pointing to null. In this case, we need to set the right-most child to point to the current node, instead of NULL and move to the left of the current node.
Case 3: When there is a left subtree and the right-most child of this left subtree is already pointing to the current node. In this case, we know that the left subtree is already visited so we need to print the current node's value and move to the right of the current node.